ANDROID: MGLRU: Don't skip anon reclaim if swap low
MGLRU tries to avoid a scan if there is not at least a full batch
(MIN_LRU_BATCH pages) that can be scanned. This can lead to a violation
of the LMKD psi poll window and cause misbehavior of the low memory
userspace killer.

Fix this by allowing anon reclaim as long as there is any swap space
left, matching the behavior of the conventional active/inactive LRU.

Bug: 268574308
Change-Id: Id381316931a9cf6e7ea8b1ea7800c77f176c9892
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>
[adapted to 4.14 FROMLIST v1 MGLRU]