BACKPORT: Multi-gen LRU: fix per-zone reclaim
MGLRU has a LRU list for each zone for each type (anon/file) in each
generation.  The min_seq (oldest generation) can progress independently
for each type but the max_seq (youngest generation) is shared for both
anon and file.  This is to maintain a common frame of reference.

In order for eviction to advance the min_seq of a type, all the
per-zone lists in the oldest generation of that type must be empty.

The eviction logic only considers pages from eligible zones for
eviction or promotion.

Consider the system has the movable zone configured and default 4
generations.  If a GFP_KERNEL allocation request comes in (eligible
zone index <= Normal), evict_pages() can return without doing any work
since there are no pages to scan in the eligible zones of the oldest
generation, even though there are reclaimable pages in newer
generations.  Reclaim won't make progress until triggered from a
ZONE_MOVABLE allocation request; which may not happen soon if there is
a lot of free memory in the movable zone.  This can lead to OOM kills.

This issue is not seen in the conventional active/inactive LRU since
there are no per-zone lists.

If there are no (not enough) pages to scan in the eligible zones, move
pages from ineligible zone (zone_index > reclaim_index) to the next
generation.  This allows for the progression of min_seq and reclaiming
from the next generation.

Qualcomm, Mediatek and raspberrypi [1] discovered this issue
independently.

[1] https://github.com/raspberrypi/linux/issues/5395

Link: https://lkml.kernel.org/r/20230802025606.346758-1-kaleshsingh@google.com
Fixes: ac35a4902374 ("mm: multi-gen LRU: minimal implementation")
Signed-off-by: Kalesh Singh <kaleshsingh@google.com>
Reported-by: Charan Teja Kalla <quic_charante@quicinc.com>
Reported-by: Lecopzer Chen <lecopzer.chen@mediatek.com>
Tested-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@collabora.com> [mediatek]
Tested-by: Charan Teja Kalla <quic_charante@quicinc.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>

(cherry picked from commit 669281ee7ef731fb5204df9d948669bf32a5e68d)
[ev: adapted from 6.5 to 4.14 FROMLIST v1 MGLRU]